CRUD 테스트

+++ src/main/java/com/example/sparta/domain/review/service/ReviewService.java
@@ -4,6 +4,9 @@ import com.example.sparta.domain.review.dto.ReviewRequestDto;
 import com.example.sparta.domain.review.dto.ReviewResponseDto;
 import com.example.sparta.domain.review.entity.Review;
 import com.example.sparta.domain.review.repository.ReviewRepository;
+import com.example.sparta.domain.store.entity.Store;
+import com.example.sparta.domain.store.repository.StoreRepository;
+import com.example.sparta.domain.user.entity.User;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +17,7 @@ import org.springframework.stereotype.Service;
 public class ReviewService {
 
     private final ReviewRepository reviewRepository;
+    private final StoreRepository storeRepository;
 
     //리뷰 전체조회
     public List<ReviewResponseDto> findAll() {
@@ -22,6 +26,7 @@ public class ReviewService {
 
     //리뷰 단건조회
     //리뷰가 존재하지 않으면 예외처리
+    //사용자 검증하기
     public ReviewResponseDto findOne(Long id) {
         Review review = reviewRepository.findById(id)
             .orElseThrow(() -> new NullPointerException("리뷰가 존재하지 않습니다."));
@@ -30,16 +35,19 @@ public class ReviewService {
     }
 
     //리뷰 수정
-    //잘못된 값 받을때 예외처리
-    //null 갑을 받을때 예외처리
-    //todo 권한 설정 (리뷰 작성자만 가능하게 구현하기)
-    public ReviewResponseDto updateOne(Long id, ReviewRequestDto reviewRequestDto) {
+    //권한 설정 (리뷰 작성자만 가능하게 구현하기)
+    public ReviewResponseDto updateOne(Long id, ReviewRequestDto reviewRequestDto, User user) {
+
 
         if (reviewRequestDto.getRating() == null) {
             throw new IllegalArgumentException("Rating 올바른 값을 넣어주세요.");
         }
         Review review = reviewRepository.findById(id)
             .orElseThrow(() -> new NullPointerException("수정할 리뷰가 없습니다."));
+
+        if(user.getUserId() == review.getReviewId()){ // 리뷰의 테이블에 있는 user_id를 가지고 와서 비교한 후에 검증한다.
+            throw new IllegalArgumentException("리뷰 수정할 수 있는 권한이없습니다.");
+        }
         review.updateOne(reviewRequestDto.getRating(), reviewRequestDto.getContent());
         Review reviewSave = reviewRepository.save(review);
 
@@ -47,9 +55,8 @@ public class ReviewService {
     }
 
     //리뷰 등록
-    //별점이 5이상 넘으면 예외처리
-    //null값을 받을때 예외처리
-    public ReviewResponseDto register(ReviewRequestDto reviewRequestDto) {
+    //todo 권한 설정(주문에대한 리뷰를 남겨야하나?)
+    public ReviewResponseDto register(ReviewRequestDto reviewRequestDto, User user) {
         if (reviewRequestDto.getRating() == null) {
             throw new NullPointerException("Rating 올바른 값을 넣어주세요.");
         }
@@ -57,18 +64,23 @@ public class ReviewService {
             throw new IllegalArgumentException("rating 6이상할 수 없습니다.");
         }
 
-        Review review = new Review(reviewRequestDto);
+        Store store = storeRepository.getReferenceById(reviewRequestDto.getStoreId());
+
+        Review review = new Review(reviewRequestDto, user, store);
+
 
         return new ReviewResponseDto(reviewRepository.save(review));
     }
 
     //리뷰 단일 삭제
-    //리뷰가 존재하지 않으면 예외처리
-    //todo 권한 체크하기
-    public void deleteOne(Long id) {
+    //리뷰에 대한 권한 체크
+    public void deleteOne(Long id, User user) {
         Review review = reviewRepository.findById(id)
             .orElseThrow(() -> new NullPointerException("존재하지 않는 리뷰입니다"));
 
+        if(review.getUser().getUserId() == user.getUserId()){
+            throw new IllegalArgumentException("리뷰를 삭제할 수 있는 권한이 없습니다.");
+        }
         reviewRepository.delete(review);
     }
 }
